+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 5.1.0 (2020-09-08)
+
+### Enhancements
+
+* Add accessor for breadcrumb list on Client
+  [#924](https://github.com/bugsnag/bugsnag-android/pull/924)
+
+* Test improvement: removed conditional operator test smell
+  [#925](https://github.com/bugsnag/bugsnag-android/pull/925)
+
+#### React Native
+
+* Keep name of class for use in reflection
+  [#927](https://github.com/bugsnag/bugsnag-android/pull/927)
+
 ## 5.0.2 (2020-08-17)
 
 ### Bug fixes